refactor: sort기능 hook으로 분리
